requirements.txt
Flask==3.1.2gunicorn==21.2.0tensorflow==2.19.0keras==3.10.0numpy==2.0.2pandas==2.2.2Pillow==10.2.0scikit-learn==1.6.1matplotlib==3.10.0seaborn==0.13.2h5py==3.15.1